+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
@@ -85,6 +86,12 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+
+        $products = Product::where(['category_id' => $category->id])->count();
+        if ($products > 0) {
+            return response(['status' => 'error', 'message' => 'This Category Have Some Products you cant Delete It.']);
+        }
+
         $category->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
